cleaned the project, changed Main class position and created a Parser
