+++ src/main/java/com/github/veluvexiau/umldoc/classFile/ClassFileParser.java
@@ -1,20 +1,19 @@
-package com.github.veluvexiau.umldoc.core;
+package com.github.veluvexiau.umldoc.classFile;
 
 import com.github.forax.umldoc.core.Entity;
 import com.github.forax.umldoc.core.Entity.Stereotype;
-import com.github.forax.umldoc.core.Field;
-import com.github.forax.umldoc.core.Method;
-import com.github.forax.umldoc.core.Modifier;
 import java.io.IOException;
 import java.lang.constant.ClassDesc;
 import java.lang.constant.MethodTypeDesc;
 import java.lang.module.ModuleFinder;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
+
+import com.github.forax.umldoc.core.Field;
+import com.github.forax.umldoc.core.Method;
+import com.github.forax.umldoc.core.Modifier;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.FieldVisitor;
@@ -22,35 +21,10 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.RecordComponentVisitor;
 
-/**
- * The main class to export UML in Mermaid and PlantUml.
- */
-public class Main {
-  /**
-   * The main method to export UML in mermaid and PlantUml.
-   */
-  public static void main(String[] args) throws IOException {
-    /* if (args.length<1){
-        throw new IllegalStateException("no path specified");
-    }*/
-    var path = Path.of("target/classes");
-    var entityFromJar = readJarFile(path);
-
-    MermaidExtract mermaid = new MermaidExtract();
-    PlantUmlExtract plantUml = new PlantUmlExtract();
-
-    mermaid.generate(entityFromJar);
-    plantUml.generate(entityFromJar);
-  }
-
-
-  private static List<Entity> readJarFile(Path path) throws IOException {
-
+public final class ClassFileParser {
+  public static List<Entity> readJarFile(Path path) throws IOException {
     var entities = new ArrayList<Entity>();
-
     var finder = ModuleFinder.of(path);
-
-
     for (var moduleReference : finder.findAll()) {
       try (var reader = moduleReference.open()) {
         for (var filename : (Iterable<String>) reader.list()::iterator) {
@@ -86,16 +60,16 @@ public class Main {
                 System.out.println(access + "here" + name);
                 if (java.lang.reflect.Modifier.isInterface(access)) {
                   entity = new Entity(Set.of(), name,
-                  Stereotype.INTERFACE, List.of(), List.of());
+                          Stereotype.INTERFACE, List.of(), List.of());
                 } else if ((access & Opcodes.ACC_RECORD) != 0) {
                   entity = new Entity(Set.of(), name,
-                  Stereotype.RECORD, List.of(), List.of());
+                          Stereotype.RECORD, List.of(), List.of());
                 } else if ((access & Opcodes.ACC_ENUM) != 0) {
                   entity = new Entity(Set.of(), name,
-                  Stereotype.ENUM, List.of(), List.of());
+                          Stereotype.ENUM, List.of(), List.of());
                 } else {
                   entity = new Entity(Set.of(), name,
-                  Stereotype.CLASS, List.of(), List.of());
+                          Stereotype.CLASS, List.of(), List.of());
                 }
                 entities.add(entity);
               }
@@ -119,21 +93,19 @@ public class Main {
                 // (like String, List<String>, Set<Entity> for exemple -->
                 // concatenation of descriptor and signature
                 if (modifier(access) != null) {
-
                   var type = ClassDesc.ofDescriptor(descriptor).displayName();
-
                   Field field = new Field(Set.of(modifier(access)), name, type);
                   var oldEntity = entities.get(entities.size() - 1);
                   var listOfFields = new ArrayList<>(oldEntity.fields());
                   listOfFields.add(field);
                   Entity entity = new Entity(oldEntity.modifiers(), oldEntity.name(),
-                      oldEntity.stereotype(), listOfFields,
-                      oldEntity.methods());
+                          oldEntity.stereotype(), listOfFields,
+                          oldEntity.methods());
                   entities.set(entities.size() - 1, entity);
                 }
                 System.out.println("  field " + modifier(access) + " " + name
-                                  + " " + ClassDesc.ofDescriptor(descriptor).displayName()
-                                  + " " + signature);
+                        + " " + ClassDesc.ofDescriptor(descriptor).displayName()
+                        + " " + signature);
 
                 return null;
               }
@@ -145,8 +117,8 @@ public class Main {
                                                String signature,
                                                String[] exceptions) {
                 System.out.println("  method " + modifier(access) + " " + name
-                                  + " " + MethodTypeDesc.ofDescriptor(desc).displayDescriptor()
-                                  + " " + signature);
+                        + " " + MethodTypeDesc.ofDescriptor(desc).displayDescriptor()
+                        + " " + signature);
                 var oldEntity = entities.get(entities.size() - 1);
                 var listOfMethods = new ArrayList<>(oldEntity.methods());
 
@@ -157,18 +129,14 @@ public class Main {
                     Method.Parameter p = new Method.Parameter(met.toString(), met.displayName());
                     parameters.add(p);
                   }
-
-
-
                   Method method = new Method(Set.of(modifier(access)), name, "" + signature,
-                      parameters);
+                          parameters);
                   listOfMethods.add(method);
                   Entity entity = new Entity(oldEntity.modifiers(), oldEntity.name(),
-                      oldEntity.stereotype(), oldEntity.fields(),
-                      listOfMethods);
+                          oldEntity.stereotype(), oldEntity.fields(),
+                          listOfMethods);
                   entities.set(entities.size() - 1, entity);
                 }
-
                 return null;
               }
             }, 0);
@@ -177,6 +145,5 @@ public class Main {
       }
     }
     return entities;
-
   }
 }
